Prevent segfault on uninitialized thread
This is a quick-and-dirty fix to prevent segmentation faults caused by
`pthread_join` on uninitialized threads when run on more than one process per
node.

I just test if the thread is initialized and skip the `pthread_join` if not.
While not being sure if this is the "clean-way", this solution is tested and
works with mutltiple tasks per node.